+++ home-configurations/core.nix
@@ -1,0 +1,26 @@
+#########################################
+# Core home configuration and packages. #
+#########################################
+{pkgs, ...}: {
+  # No-configuration packages.
+  home.packages = with pkgs; [
+    # Core packages.
+    gnupg
+
+    # Programming languages (or related).
+    rustup
+    typst
+    texliveFull
+
+    # Python-related packages.
+    uv
+    poetry
+
+    # Miscellaneous.
+    cmatrix
+  ];
+
+  # Minimal configuration-dependent programs and/or services.
+  programs = {};
+  services = {};
+}