+++ system-configs.nix
@@ -5,8 +5,10 @@ inputs @ {
   nixpkgs,
   nix-darwin,
   home-manager,
+  catppuccin,
   ...
 }: let
+  # The custom 'configurations-manager' library.
   configurations-manager = import ./configurations-manager {
     nixpkgs-lib = nixpkgs.lib;
     inherit
@@ -17,9 +19,13 @@ inputs @ {
     rootDir = ./.;
   };
 in {
+  # Configurations for Darwin systems.
   darwinConfigurations = configurations-manager.mkDarwinConfigurations {
-    specialArgs = inputs;
+    specialArgs = inputs // {computerName = "Mario's MacBook Pro";};
     homeSpecialArgs = inputs;
+    extraHomeSharedModules = [
+      catppuccin.homeManagerModules.catppuccin
+    ];
     extraHomeManagerArgs = {
       backupFileExtension = "backup";
       useGlobalPkgs = true;
@@ -32,6 +38,7 @@ in {
         users = {
           "mariovagomarzal" = {
             userInfo = {
+              userName = "Mario Vago Marzal";
               email = "mariovagomarzal@gmail.com";
               githubUser = "mariovagomarzal";
             };