+++ CHANGELOG.md
@@ -18,6 +18,31 @@ sections:
   `README.md` or `CHANGELOG.md`.
 - _Miscellaneous_: for changes that do not fit in any of the previous sections.
 
+## [2025.01.11]
+
+### Mario's MacBook Pro
+
+- The behavior of some Darwin shared modules is now more general.
+- Many system settings options have been added.
+- Packages installation options have been reorganized. Now, 'nix-darwin' manages
+  less packages in favor of 'home-manager'.
+- Now 'home-manager' programs/services use the Catppuccin theme.
+- Added the following programs/services for 'home-manager':
+  - Bat
+  - LSD
+  - Visual Studio Code
+  - Firefox
+
+### Development
+
+- Some 'Markdownlint' rules have been disabled.
+
+### Documentation
+
+- A logo and some badges have been added to the README file.
+- The documentation of the setup process for Mario's MacBook Pro has been
+  improved.s
+
 ## [2024.12.28]
 
 ### Codebase
@@ -85,6 +110,8 @@ sections:
   [dotbot](https://github.com/anishathalye/dotbot) tool.
 
 <!-- External links -->
+[2025.01.11]:
+  https://github.com/mariovagomarzal/dotfiles/releases/tag/2025.01.11
 [2024.12.28]:
   https://github.com/mariovagomarzal/dotfiles/releases/tag/2024.12.28
 [2024.12.12]: